Move plugin settings to individual plugin classes
