+++ web/app/themes/wordpress-scaffold/lib/Grrr/Utils/AdminEditor.php
@@ -84,3 +84,21 @@ function user_can_richedit_custom() {
     return false;
 }
 add_filter('user_can_richedit', __NAMESPACE__ . '\\user_can_richedit_custom');
+
+/**
+ * Set default image embed options for WYSIWYG.
+ * See: https://core.trac.wordpress.org/ticket/35101
+ */
+function reset_image_insert_settings() {
+  ?>
+    <script>
+      if (typeof setUserSetting !== 'undefined') {
+        setUserSetting('imgsize', 'large'); // thumbnail || medium || large || full
+        setUserSetting('align', 'none'); // none || left || center || right
+        setUserSetting('urlbutton', 'none'); // none || file || post
+      }
+    </script>
+  <?php
+}
+add_action('admin_head-post.php', __NAMESPACE__ . '\\reset_image_insert_settings');
+add_action('admin_head-post-new.php', __NAMESPACE__ . '\\reset_image_insert_settings');
