+++ web/app/themes/wordpress-scaffold/lib/Grrr/Utils/Extras.php
@@ -51,32 +51,6 @@ function set_max_srcset(int $max, array $sizes) {
 add_filter('max_srcset_image_width', __NAMESPACE__ . '\\set_max_srcset', 10, 2);
 
 /**
- * Set the Open Graph image size (else will use native 'Large' from WordPress).
- */
-function yoast_og_size() {
-    return 'image--huge';
-}
-add_filter('wpseo_opengraph_image_size', __NAMESPACE__ . '\\yoast_og_size', 10, 2);
-
-/**
- * Set default image embed options for WYSIWYG.
- * See: https://core.trac.wordpress.org/ticket/35101
- */
-function reset_image_insert_settings() {
-  ?>
-    <script>
-      if ( typeof setUserSetting !== 'undefined' ) {
-        setUserSetting('imgsize', 'large'); // thumbnail || medium || large || full
-        setUserSetting('align', 'none'); // none || left || center || right
-        setUserSetting('urlbutton', 'none'); // none || file || post
-      }
-    </script>
-  <?php
-}
-add_action('admin_head-post.php', __NAMESPACE__ . '\\reset_image_insert_settings');
-add_action('admin_head-post-new.php', __NAMESPACE__ . '\\reset_image_insert_settings');
-
-/**
  * Adjust embed size defaults.
  */
 function embed_defaults($embed_size){
